Crate literate

Modules

walkwalk

Structs

LanguageMatcher

Matches the fenced code block against an Optional language. Exposes control over whether to include fenced code blocks without a language in the output.

Enums

LiterateError

Traits

CodeMatcher

Determines whether a fenced code block should be included in the output.

Functions

extract

Extracts fenced code blocks from the input. If the matcher returns true for the language, the contents of the code block are written to the output. Otherwise, the contents are ignored. If the language of the fenced code block is blank (empty or blank space), the matcher will get None as the language.