Function literate::extract

pub fn extract<Input, Output, Matcher>(
    input: Input, 
    output: Output, 
    matcher: Matcher
) -> Result<usize, LiterateError> where
    Input: Read,
    Output: Write,
    Matcher: CodeMatcher, 

Extracts fenced code blocks from the input. If the matcher returns true for the language, the contents of the code block are written to the output. Otherwise, the contents are ignored. If the language of the fenced code block is blank (empty or blank space), the matcher will get None as the language.

Returns the number of extracted bytes.