Function literate::walk::extract

pub fn extract<Input, Filter, Matcher, Mapper>(
    input: Input, 
    filter: Filter, 
    mapper: Mapper, 
    matcher: Matcher, 
    overwrite: bool
) -> Result<usize, LiterateError> where
    Input: AsRef<Path>,
    Filter: FileFilter,
    Matcher: CodeMatcher,
    Mapper: PathMapper, 

This is supported on crate feature walk only.

Walks a the directory tree and extracts matching files. The resulting content is written to the mapped filename.

Returns the number of extracted files.