Trait literate::walk::FileFilter

pub trait FileFilter {
    // Required method
    fn filter_file(&self, file: impl AsRef<Path>) -> bool;
}

Available on crate feature walk only.

Determines whether a file will be included in the directory tree walking extraction or not.

Required Methods

fn filter_file(&self, file: impl AsRef<Path>) -> bool

Tests whether this file should be included in the list of files to extract.

Object Safety

This trait is not object safe.